default total vol hours report technically works but very jank and some non-critical errors
Gives total volunteer hours but only if no fields are filled in (name, name range, date range, status).

There are also quite a few changes and comments that can probably be removed, so the differences between this version and the previous one should be reviewed and gone through. but I am tired and want to go to sleep so I'm committing this because my coding session is over.