Updated Remove Course, Update Course, and Calendar.
Update Course and Remove Course now have a search bar so in the case there are several courses, the user can easily find the one they're looking for. Updated the calendar to display the name of the course (the only required field), rather than the abbreviated course name as that field can now be left empty. Added Manage Courses to the header for easier access.